wgpu-core: Move storage code to new `storage` module.
