+++ wgpu-core/src/command/bundle.rs
@@ -91,12 +91,13 @@ use crate::{
     },
     error::{ErrorFormatter, PrettyError},
     hal_api::HalApi,
-    hub::{Hub, Resource, Storage, Token},
+    hub::{Hub, Resource, Token},
     id,
     identity::GlobalIdentityHandlerFactory,
     init_tracker::{BufferInitTrackerAction, MemoryInitKind, TextureInitTrackerAction},
     pipeline::{self, PipelineFlags},
     resource,
+    storage::Storage,
     track::RenderBundleScope,
     validation::check_buffer_usage,
     Label, LabelHelpers, LifeGuard, Stored,
