+++ wgpu-core/src/command/memory_init.rs
@@ -5,10 +5,10 @@ use hal::CommandEncoder;
 use crate::{
     device::Device,
     hal_api::HalApi,
-    hub::Storage,
     id::{self, TextureId},
     init_tracker::*,
     resource::{Buffer, Texture},
+    storage::Storage,
     track::{TextureTracker, Tracker},
     FastHashMap,
 };
